[API] Rename directories
